Refactored `controller` module
Split the controller module into a number of more specific modules.
Making navigation easier, will also enable programmatic code
generation to be implemented in a simpler manner.
